@@ -9,13 +9,14 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@emotion/cache": "^11.10.5",
-    "@emotion/react": "^11.10.6",
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.11.4",
     "@emotion/server": "^11.10.0",
-    "@emotion/styled": "^11.10.6",
+    "@emotion/styled": "^11.11.5",
     "@mui/icons-material": "^5.11.11",
     "@mui/lab": "^5.0.0-alpha.124",
-    "@mui/material": "^5.11.15",
+    "@mui/material": "^5.16.1",
+    "@mui/material-nextjs": "^5.16.1",
     "@mui/x-date-pickers": "^6.11.1",
     "@tanstack/react-query": "^5.51.1",
     "@tanstack/react-query-devtools": "^5.51.1",
@@ -26,9 +27,9 @@
     "date-fns": "^2.29.3",
     "formik": "^2.2.9",
     "jwt-decode": "^3.1.2",
-    "next": "13.2.4",
+    "next": "^14.2.5",
     "nprogress": "^0.2.0",
-    "react": "18.2.0",
+    "react": "^18.3.1",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
     "react-toastify": "^9.1.2",
@@ -36,6 +37,7 @@
     "remark-html": "^15.0.2",
     "simplebar": "^6.2.4",
     "simplebar-react": "^3.2.3",
+    "styled-components": "^6.1.11",
     "typescript": "5.0.3",
     "yup": "^1.0.2"
   },
