@@ -1,5 +1,5 @@
-import type { Collection } from "@/types/collection";
-import { appFetch } from "@/utils/app-fetch";
+import type { Collection } from "../app/types/collection";
+import { appFetch } from "../app/utils/app-fetch";
 import { useMutation } from "@tanstack/react-query";
 
 export const useCreateCollection = () => useMutation<any, Error, BodyInit>({
