@@ -1,5 +1,5 @@
-import { Developer } from "@/types/developer";
-import { appFetch } from "@/utils/app-fetch";
+import { Developer } from "../app/types/developer";
+import { appFetch } from "../app/utils/app-fetch";
 import { useMutation } from "@tanstack/react-query";
 
 export const useCreateDeveloper = (onSuccess: any) =>
